+++ tests/cases/ep_svc/service.cpp
@@ -24,10 +24,10 @@ void skull_service_release(skullcpp::Service& service)
 }
 
 static
-size_t _ep_unpack(const void* data, size_t len)
+ssize_t _ep_unpack(const void* data, size_t len)
 {
     std::cout << "ep received data len: " << len << std::endl;
-    return len;
+    return (ssize_t)len;
 }
 
 static
